+++ src/main/java/net/rgsw/noise/Perlin3D.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 RGSW
+ * Copyright (c) 2020 RGSW
  * Licensed under Apache 2.0 license
  */
 
@@ -78,7 +78,7 @@ public class Perlin3D extends Noise3D {
         { 1, 0, 1 }
     };
 
-    private int gradIndex( int x, int y, int z ) {
+    private int gradIndex( long x, long y, long z ) {
         int hash = Hash.hash3I( seed, x, y, z );
         return hash & 31;
     }
@@ -91,9 +91,9 @@ public class Perlin3D extends Noise3D {
         return t * t * t * ( t * ( t * 6 - 15 ) + 10 );
     }
 
-    private static int fastfloor( double v ) {
+    private static long fastfloor( double v ) {
         int d = v < 0 ? - 1 : 0;
-        return (int) v + d;
+        return (long) v + d;
     }
 
     @Override
@@ -102,12 +102,12 @@ public class Perlin3D extends Noise3D {
         y /= scaleY;
         z /= scaleZ;
 
-        int minx = fastfloor( x );
-        int miny = fastfloor( y );
-        int minz = fastfloor( z );
-        int maxx = minx + 1;
-        int maxy = miny + 1;
-        int maxz = minz + 1;
+        long minx = fastfloor( x );
+        long miny = fastfloor( y );
+        long minz = fastfloor( z );
+        long maxx = minx + 1L;
+        long maxy = miny + 1L;
+        long maxz = minz + 1L;
         double smoothx = smooth( x - minx );
         double smoothy = smooth( y - miny );
         double smoothz = smooth( z - minz );
